Add support for SQLite (#37)
Implement support for SQLite, along-side MySQL/MariaDB. While an existing
Firefly-III installation is likely to be running on MySQL, SQLite is a simpler
and lighter alternative if lychnos is running on a different host. Since we're
also automatically creating the database file if none is specified, this
simplifies setup for new users. Eventually, I would like to be able to
distribute lychnos in a Docker image, and removing the dependency on MySQL is a
first step to accomplish this.

Caveat: personally I'm using MySQL, so SQLite is not yet thoroughly tested.