Periodically check if caches should be invalidated (#279)
When I edit or input a transaction in Firefly-III, lychnos doesn't know about
it. To keep the lychnos cache fresh, check the asset account balances in Firefly
every minute (the request takes about 150ms for me). If any of the balances
changed, then invalidate and refresh all the caches.

Future idea: it could be useful if the refresh interval were configurable.